requirements.txt
absl-py==0.2.2
astor==0.6.2
astroid==2.2.5
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
bleach==1.5.0
blinker==1.3
bs4==0.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==2.3.0
Click==7.0
cloud-init==18.5
command-not-found==0.3
configobj==5.0.6
coverage==4.5.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.2
django-contrib-comments==1.9.1
django-extensions==2.2.1
django-import-export==1.2.0
et-xmlfile==1.0.1
Flask==1.0.2
Frozen-Flask==0.15
gast==0.2.0
grpcio==1.12.1
gunicorn==19.9.0
html5lib==0.9999999
idna==2.0
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.16
jdcal==1.4.1
jedi==0.12.0
Jinja2==2.10.1
jsonpatch==1.10
jsonpointer==1.9
kiwisolver==1.0.1
language-selector==0.1
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
mccabe==0.6.1
numpy==1.14.5
oauthlib==1.0.3
odfpy==1.4.0
openpyxl==2.6.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
pipenv==2018.10.13
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.5.2
pyasn1==0.1.9
pycparser==2.18
pycurl==7.43.0
Pygments==2.2.0
PyJWT==1.3.0
pylint==2.3.1
pyparsing==2.2.0
pyserial==3.0.1
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.11
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
sounddevice==0.3.11
soupsieve==1.9.1
tablib==0.13.0
termcolor==1.1.0
traitlets==4.3.2
typed-ast==1.3.1
ufw==0.35
urllib3==1.13.1
virtualenv==16.0.0
Wave==0.0.2
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
